Update Freetype to 2.14.2
[ChangeLog][Third-Party Code] Updated bundled Freetype to
2.14.2.

Pick-to: 6.11.0 6.11 6.10 6.10.3 6.8 6.5 5.15
Fixes: QTBUG-144722

+++ src/3rdparty/freetype/src/tools/apinames.c
@@ -101,7 +101,7 @@ names_add( const char*   name,
   {
     nm = the_names + nn;
 
-    if ( (int)nm->hash                 == h &&
+    if ( nm->hash                      == h &&
          memcmp( name, nm->name, len ) == 0 &&
          nm->name[len]                 == 0 )
       return;
